Initial POC of pvc transfer
This POC just uses annotations to prove the idea, here's what happens:

1. User adds an annotation to request transfer of their bound pvc
2. PV claim sync detects this update, and updates an annotation on the
PV marking its claim as pending transfer
3. User deletes PVC, the PV controller checks for this annotation when
cleaning up, if it's there, it releases the PV and sets a claim ref for
the destination NS and PVC name
4. Receiving user issues a request using the PVC name specified in the
transfer which will utilize the claimref we set.

Things to improve:
1. Make globally configurable (enable/disable with a default of disabled)
2. Add paramters to PVC object rather than using annotations for request
3. Add a mechanism to avoid any accidental import of volumes (ie another
spec entry that only matches trasnferred volumes if they match)?
4. Determine if we need our own controller rather than using the exising
PV controller.. and if so can we give it permissions to modify PVs
safely?

If you want to try it out here are the steps you can use:
1.  Create a pvc just like you normally would
2.  Upate the pvc with a transfer annotation:
    ```<namespace>/<new-pvc-name>```
    NOTE you can do this in the same namespace too if you like
         Name MUST differ!!
3.  Delete the PVC, the PV should still exist with a
    transfer-status=pending annotation and a claim for the new
    namespace/name
4.  Request a new PVC with the name/namespace you specified in the transfer
    annotation, upon success transfer-status goes to complete
5.  Delete the PVC and the volume will now be deleted as well